backend/requirements.txt
Django==4.1.13
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.3 